Scripts e Snips/CyberSecurity/Ransoware/requirements.txt
requests==2.28.2
cryptography==40.0.2
python-dotenv==1.0.0
